Fix race condition in tests
When we send an event to riemann, it is not immediately query-able from
the index.  If a previously matching event was found, it was used
instead of waiting for the new actual event we where waiting for,
causing a failing test in CI.

Because tests are randomized and because the amount of time CPU given to
each process change from one run to another, these failures where not
100% percent reproducible, but are seen from time to time. For example,
this PR so far has experienced 4 failures across the 30 runs of the
previous commits.

Setup a singleton class that manage a monotonic sequence of numbers, and
use these numbers to build a "message_id" attribute.  Always use this
attribute to get an event rather than its service, state, metric, or any
other attribute.

This should fix these transient CI failures.