Allow TLS 1.3
`SSLContext#ssl_version` is deprecated and only provided for backwards
compatibility. Using `SSLContext#min_version` and
`SSLContext#max_version` give a beter control on the version of TLS used
by the client.

Set TLS 1.2 as the minimum TLS version required for establishing the
connexion.  This allow riemann-ruby-client to connect to Riemann using
TLS 1.3.